+++ MobileApp/MobileApp/MobileApp/User.cs
@@ -10,7 +10,7 @@ namespace MobileApp
         public UserName Name { get; set; }
         public double Plan { get; set; }
         public bool AdminStatus { get; set; }
-        public int Used { get; set; }
+        public double Used { get; set; }
         public double Allocated { get; set; }
         public string PlanStartDate { get; set; }
         public string PlanEndDate { get; set; }
